% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_datasus.R
\name{fetch_datasus}
\alias{fetch_datasus}
\title{Fetch and read microdata files from DataSUS}
\usage{
fetch_datasus(
  year_start,
  month_start = NULL,
  year_end,
  month_end = NULL,
  uf = "all",
  information_system,
  vars = NULL,
  stop_on_error = FALSE,
  timeout = 240,
  track_source = FALSE
)
}
\value{
a \code{data.frame} with the contents of the DBC files.
}
\description{
\code{fetch_datasus} downloads microdata (DBC) files from DataSUS and reads them.
}
\section{Fetch two years of data from SIM-DO}{
res <- fetch_datasus(year_start = 2010, year_end = 2011, uf = "AC",
information_system = "SIM-DO")
}

\section{Fetch one year of data from SIM-DO and keep only three variables}{
res <- fetch_datasus(year_start = 2014, year_end = 2014,
information_system = "SIM-DO", uf = "AC",
vars = c("CODMUNRES", "DTOBITO", "CAUSABAS"))
}

\section{Fetch some months' data from SIH-RD for four states}{
res <- fetch_datasus(year_start = 2014, month_start = 1,
year_end = 2014, month_end = 2,
uf = c("AC", "RR"),
information_system = "SIH-RD")
}

