% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sih_fiocruz.R
\name{fetch_sih_fiocruz}
\alias{fetch_sih_fiocruz}
\title{Fetch SIH Data from DataSUS by State (Fiocruz)}
\usage{
fetch_sih_fiocruz(
  year,
  month,
  uf = "all",
  timeout = 240,
  stop_on_error = FALSE,
  track_source = FALSE
)
}
\arguments{
\item{year}{Numeric value indicating the year. Must be 1996 or later.}

\item{month}{Numeric value indicating the month. Must be between 1 and 12.}

\item{uf}{Character vector of state codes representing the federative units. Default is \code{"all"} (fetches data for all states).}

\item{timeout}{Numeric value indicating the maximum number of seconds to wait for each download. Default is 240 seconds.}

\item{stop_on_error}{Logical value. If \code{TRUE}, stops the function if any download fails. Default is \code{FALSE}.}

\item{track_source}{Logical value. If \code{TRUE}, adds a \code{source} column to the data, tracking the original file for each row. Default is \code{FALSE}.}
}
\value{
A data frame with the downloaded and merged data from DataSUS, or \code{NULL} if the FTP server is down or there is no internet connection.
}
\description{
This function fetches and merges SIH (Hospital Information System) data from DataSUS for a specified list of states or for all states by default.
It downloads \code{.dbc} files from DataSUS's FTP server, processes them one by one, and returns a combined dataset.
}
\examples{
\dontrun{
# Fetch SIH data for January 2021 for all states
data <- fetch_sih_fiocruz(year = 2021, month = 1)

# Fetch SIH data for January 2021 for a specific list of states
data <- fetch_sih_fiocruz(year = 2021, month = 1, uf = c("SP", "MG", "RJ"))
}
}
