% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_codes.R
\docType{data}
\name{icd_codes}
\alias{icd_codes}
\title{ICD Codes Dataset}
\format{
A data frame with several rows and columns:
\describe{
\item{code}{ICD code}
\item{Description}{Description of the ICD code}
}
}
\source{
\url{Your data source or reference here}
}
\usage{
icd_codes
}
\description{
A dataset containing ICD codes and their descriptions.
}
\keyword{datasets}
