% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdatasus.R
\docType{package}
\name{microdatasus}
\alias{phmm2000}
\alias{phmm2000-package}
\alias{microdatasus}
\title{microdatasus: Download and preprocess DataSUS files}
\description{
Downloads microdata files (DBC format) from DataSUS and imports the files for use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PauloHenriqueMelo/test}
  \item Report bugs at \url{https://github.com/rfsaldanha/microdatasus/issues}
}

}
\author{
\strong{Maintainer}: Raphael Saldanha \email{raphael.saldanha@fiocruz.br} (\href{https://orcid.org/0000-0003-0652-8466}{ORCID})

}
