% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgsscdata.R
\docType{package}
\name{pgsscdata}
\alias{teambrazil}
\alias{teambrazil-package}
\alias{pgsscdata}
\title{pgssc_data: Download and preprocess DataSUS files}
\description{
Downloads microdata files (DBC format) from DataSUS and imports the files for use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PauloHenriqueMelo/teambrazil}
  \item Report bugs at \url{phmm2000@gmail.com}
}

}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Paulo Melo \email{phmm2000@gmail.com} (\href{https://orcid.org/0000-0001-6147-5833}{ORCID})
}

}
