% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rea.R
\docType{data}
\name{proc_rea}
\alias{proc_rea}
\title{A dataset containing SigTap DATASUS procedure codes}
\format{
A data frame with several rows and columns:
\describe{
\item{code}{SigTap code}
\item{PROCEDURE}{Description of the ICD code}
}
}
\source{
\url{Your data source or reference here}
}
\usage{
proc_rea
}
\description{
A dataset containing SigTap DATASUS procedure codes
}
\keyword{datasets}
