% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rd.R
\name{process_rd}
\alias{process_rd}
\title{This function processes key SIH variables, converting them to appropriate formats and labels, if the variables exist in the data.}
\usage{
process_rd(data)
}
\arguments{
\item{data}{A \code{data.frame} containing SIH data.}
}
\value{
A \code{data.frame} with the processed data.
}
\description{
This function processes key SIH variables, converting them to appropriate formats and labels, if the variables exist in the data.
}
\examples{
# Example usage:
process_sih_rd(sih_rd_sample)
}
